Refactor - move thread dispatching and logging/overload out of the stack into separate modules
