+++ sprout/stack.cpp
@@ -79,24 +79,6 @@ class StackQuiesceHandler;
 
 struct stack_data_struct stack_data;
 
-static std::vector<pj_thread_t*> pjsip_threads;
-static std::vector<pj_thread_t*> worker_threads;
-static volatile pj_bool_t quit_flag;
-
-// Queue for incoming messages.
-struct rx_msg_qe
-{
-  pjsip_rx_data* rdata;    // received message
-  Utils::StopWatch stop_watch;    // stop watch for tracking message latency
-};
-eventq<struct rx_msg_qe> rx_msg_q;
-
-// Deadlock detection threshold for the message queue (in milliseconds).  This
-// is set to roughly twice the expected maximum service time for each message
-// (currently four seconds, allowing for four Homestead/Homer interactions
-// from a single request, each with a possible 500ms timeout).
-static const int MSG_Q_DEADLOCK_TIME = 4000;
-
 static Accumulator* latency_accumulator;
 static Accumulator* queue_size_accumulator;
 static Counter* requests_counter;
@@ -107,27 +89,8 @@ static QuiescingManager *quiescing_mgr = NULL;
 static StackQuiesceHandler *stack_quiesce_handler = NULL;
 static ConnectionTracker *connection_tracker = NULL;
 
-// We register a single module to handle scheduling plus local and
-// SAS logging.
-static pj_bool_t on_rx_msg(pjsip_rx_data* rdata);
-static pj_status_t on_tx_msg(pjsip_tx_data* tdata);
-
-static pjsip_module mod_stack =
-{
-  NULL, NULL,                           /* prev, next.          */
-  pj_str("mod-stack"),                  /* Name.                */
-  -1,                                   /* Id                   */
-  PJSIP_MOD_PRIORITY_TRANSPORT_LAYER-1, /* Priority             */
-  NULL,                                 /* load()               */
-  NULL,                                 /* start()              */
-  NULL,                                 /* stop()               */
-  NULL,                                 /* unload()             */
-  &on_rx_msg,                           /* on_rx_request()      */
-  &on_rx_msg,                           /* on_rx_response()     */
-  &on_tx_msg,                           /* on_tx_request()      */
-  &on_tx_msg,                           /* on_tx_response()     */
-  NULL,                                 /* on_tsx_state()       */
-};
+static volatile pj_bool_t quit_flag;
+static pj_thread_t* pjsip_thread;
 
 const static std::string _known_statnames[] = {
   "client_count",
@@ -161,7 +124,7 @@ const int num_known_stats = sizeof(_known_statnames) / sizeof(std::string);
 
 /// PJSIP threads are donated to PJSIP to handle receiving at transport level
 /// and timers.
-static int pjsip_thread(void *p)
+static int pjsip_thread_func(void *p)
 {
   pj_time_val delay = {0, 10};
 
@@ -178,312 +141,6 @@ static int pjsip_thread(void *p)
 
   return 0;
 }
-
-
-/// Worker threads handle most SIP message processing.
-static int worker_thread(void* p)
-{
-  // Set up data to always process incoming messages at the first PJSIP
-  // module after our module.
-  pjsip_process_rdata_param rp;
-  pjsip_process_rdata_param_default(&rp);
-  rp.start_mod = &mod_stack;
-  rp.idx_after_start = 1;
-
-  LOG_DEBUG("Worker thread started");
-
-  struct rx_msg_qe qe = {0};
-
-  while (rx_msg_q.pop(qe))
-  {
-    pjsip_rx_data* rdata = qe.rdata;
-    if (rdata)
-    {
-      LOG_DEBUG("Worker thread dequeue message %p", rdata);
-      pjsip_endpt_process_rx_data(stack_data.endpt, rdata, &rp, NULL);
-      LOG_DEBUG("Worker thread completed processing message %p", rdata);
-      pjsip_rx_data_free_cloned(rdata);
-
-      unsigned long latency_us;
-      if (qe.stop_watch.read(latency_us))
-      {
-        LOG_DEBUG("Request latency = %ldus", latency_us);
-        latency_accumulator->accumulate(latency_us);
-        load_monitor->request_complete(latency_us);
-      }
-      else
-      {
-        LOG_ERROR("Failed to get done timestamp: %s", strerror(errno));
-      }
-    }
-  }
-
-  LOG_DEBUG("Worker thread ended");
-
-  return 0;
-}
-
-
-static void local_log_rx_msg(pjsip_rx_data* rdata)
-{
-  LOG_VERBOSE("RX %d bytes %s from %s %s:%d:\n"
-              "--start msg--\n\n"
-              "%.*s\n"
-              "--end msg--",
-              rdata->msg_info.len,
-              pjsip_rx_data_get_info(rdata),
-              rdata->tp_info.transport->type_name,
-              rdata->pkt_info.src_name,
-              rdata->pkt_info.src_port,
-              (int)rdata->msg_info.len,
-              rdata->msg_info.msg_buf);
-}
-
-
-static void local_log_tx_msg(pjsip_tx_data* tdata)
-{
-  LOG_VERBOSE("TX %d bytes %s to %s %s:%d:\n"
-              "--start msg--\n\n"
-              "%.*s\n"
-              "--end msg--",
-              (tdata->buf.cur - tdata->buf.start),
-              pjsip_tx_data_get_info(tdata),
-              tdata->tp_info.transport->type_name,
-              tdata->tp_info.dst_name,
-              tdata->tp_info.dst_port,
-              (int)(tdata->buf.cur - tdata->buf.start),
-              tdata->buf.start);
-}
-
-
-static void sas_log_rx_msg(pjsip_rx_data* rdata)
-{
-  SAS::TrailId trail = 0;
-
-  if (rdata->msg_info.msg->type == PJSIP_RESPONSE_MSG)
-  {
-    // Message is a response, so try to correlate to an existing UAC
-    // transaction using the top-most Via header.
-    pj_str_t key;
-    pjsip_tsx_create_key(rdata->tp_info.pool, &key, PJSIP_ROLE_UAC,
-                         &rdata->msg_info.cseq->method, rdata);
-    pjsip_transaction* tsx = pjsip_tsx_layer_find_tsx(&key, PJ_TRUE);
-    if (tsx)
-    {
-      // Found the UAC transaction, so get the trail if there is one.
-      trail = get_trail(tsx);
-
-      // Unlock tsx because it is locked in find_tsx()
-      pj_grp_lock_release(tsx->grp_lock);
-    }
-  }
-  else if (rdata->msg_info.msg->line.req.method.id == PJSIP_ACK_METHOD)
-  {
-    // Message is an ACK, so try to correlate it to the existing UAS
-    // transaction using the top-most Via header.
-    pj_str_t key;
-    pjsip_tsx_create_key(rdata->tp_info.pool, &key, PJSIP_UAS_ROLE,
-                         &rdata->msg_info.cseq->method, rdata);
-    pjsip_transaction* tsx = pjsip_tsx_layer_find_tsx(&key, PJ_TRUE);
-    if (tsx)
-    {
-      // Found the UAS transaction, so get the trail if there is one.
-      trail = get_trail(tsx);
-
-      // Unlock tsx because it is locked in find_tsx()
-      pj_grp_lock_release(tsx->grp_lock);
-    }
-  }
-  else if (rdata->msg_info.msg->line.req.method.id == PJSIP_CANCEL_METHOD)
-  {
-    // Message is a CANCEL request chasing an INVITE, so we want to try to
-    // correlate it to the INVITE trail for the purposes of SAS tracing.
-    pj_str_t key;
-    pjsip_tsx_create_key(rdata->tp_info.pool, &key, PJSIP_UAS_ROLE,
-                         pjsip_get_invite_method(), rdata);
-    pjsip_transaction* tsx = pjsip_tsx_layer_find_tsx(&key, PJ_TRUE);
-    if (tsx)
-    {
-      // Found the INVITE UAS transaction, so get the trail if there is one.
-      trail = get_trail(tsx);
-
-      // Unlock tsx because it is locked in find_tsx()
-      pj_grp_lock_release(tsx->grp_lock);
-    }
-  }
-
-  if (trail == 0)
-  {
-    // The message doesn't correlate to an existing trail, so create a new
-    // one.
-    trail = SAS::new_trail(1u);
-  }
-
-  // Store the trail in the message as it gets passed up the stack.
-  set_trail(rdata, trail);
-
-  // Log the message event.
-  SAS::Event event(trail, SASEvent::RX_SIP_MSG, 0);
-  event.add_static_param(pjsip_transport_get_type_from_flag(rdata->tp_info.transport->flag));
-  event.add_static_param(rdata->pkt_info.src_port);
-  event.add_var_param(rdata->pkt_info.src_name);
-  event.add_compressed_param(rdata->msg_info.len, rdata->msg_info.msg_buf, &SASEvent::PROFILE_SIP);
-  SAS::report_event(event);
-}
-
-
-static void sas_log_tx_msg(pjsip_tx_data *tdata)
-{
-  // For outgoing messages always use the trail identified in the module data
-  SAS::TrailId trail = get_trail(tdata);
-
-  if (trail != 0)
-  {
-    // Log the message event.
-    SAS::Event event(trail, SASEvent::TX_SIP_MSG, 0);
-    event.add_static_param(pjsip_transport_get_type_from_flag(tdata->tp_info.transport->flag));
-    event.add_static_param(tdata->tp_info.dst_port);
-    event.add_var_param(tdata->tp_info.dst_name);
-    event.add_compressed_param((int)(tdata->buf.cur - tdata->buf.start),
-                               tdata->buf.start,
-                               &SASEvent::PROFILE_SIP);
-    SAS::report_event(event);
-  }
-  else
-  {
-    LOG_ERROR("Transmitting message with no SAS trail identifier\n%.*s",
-              (int)(tdata->buf.cur - tdata->buf.start),
-              tdata->buf.start);
-  }
-}
-
-
-static pj_bool_t on_rx_msg(pjsip_rx_data* rdata)
-{
-  // Do logging.
-  local_log_rx_msg(rdata);
-  sas_log_rx_msg(rdata);
-
-  requests_counter->increment();
-
-  // Check whether the request should be processed
-  if (!(load_monitor->admit_request()) &&
-      (rdata->msg_info.msg->type == PJSIP_REQUEST_MSG) &&
-      (rdata->msg_info.msg->line.req.method.id != PJSIP_ACK_METHOD))
-  {
-    // Discard non-ACK requests if there are no available tokens.
-    // Respond statelessly with a 503 Service Unavailable, including a
-    // Retry-After header with a zero length timeout.
-    LOG_DEBUG("Rejected request due to overload");
-
-    pjsip_cid_hdr* cid = (pjsip_cid_hdr*)rdata->msg_info.cid;
-
-    SAS::TrailId trail = get_trail(rdata);
-
-    SAS::Marker start_marker(trail, MARKER_ID_START, 1u);
-    SAS::report_marker(start_marker);
-
-    SAS::Event event(trail, SASEvent::SIP_OVERLOAD, 0);
-    event.add_static_param(load_monitor->get_target_latency());
-    event.add_static_param(load_monitor->get_current_latency());
-    event.add_static_param(load_monitor->get_rate_limit());
-    SAS::report_event(event);
-
-    PJUtils::report_sas_to_from_markers(trail, rdata->msg_info.msg);
-
-    if ((rdata->msg_info.msg->line.req.method.id == PJSIP_REGISTER_METHOD) ||
-        ((pjsip_method_cmp(&rdata->msg_info.msg->line.req.method, pjsip_get_subscribe_method())) == 0) ||
-        ((pjsip_method_cmp(&rdata->msg_info.msg->line.req.method, pjsip_get_notify_method())) == 0))
-    {
-      // Omit the Call-ID for these requests, as the same Call-ID can be
-      // reused over a long period of time and produce huge SAS trails.
-      PJUtils::mark_sas_call_branch_ids(trail, NULL, rdata->msg_info.msg);
-    }
-    else
-    {
-      PJUtils::mark_sas_call_branch_ids(trail, cid, rdata->msg_info.msg);
-    }
-
-    SAS::Marker end_marker(trail, MARKER_ID_END, 1u);
-    SAS::report_marker(end_marker);
-
-    pjsip_retry_after_hdr* retry_after = pjsip_retry_after_hdr_create(rdata->tp_info.pool, 0);
-    PJUtils::respond_stateless(stack_data.endpt,
-                               rdata,
-                               PJSIP_SC_SERVICE_UNAVAILABLE,
-                               NULL,
-                               (pjsip_hdr*)retry_after,
-                               NULL);
-
-    // We no longer terminate TCP connections on overload as the shutdown has
-    // to wait for existing transactions to end and therefore it takes too
-    // long to get feedback to the downstream node.  We expect downstream nodes
-    // to rebalance load if possible triggered by receipt of the 503 responses.
-
-    overload_counter->increment();
-    return PJ_TRUE;
-  }
-
-  // Check that the worker threads are not all deadlocked.
-  if (rx_msg_q.is_deadlocked())
-  {
-    // The queue has not been serviced for sufficiently long to imply that
-    // all the worker threads are deadlock, so exit the process so it will be
-    // restarted.
-    LOG_ERROR("Detected worker thread deadlock - exiting");
-    abort();
-  }
-
-  // Before we start, get a timestamp.  This will track the time from
-  // receiving a message to forwarding it on (or rejecting it).
-  struct rx_msg_qe qe;
-  qe.stop_watch.start();
-
-  // Notify the connection tracker that the transport is active.
-  connection_tracker->connection_active(rdata->tp_info.transport);
-
-  // Clone the message and queue it to a scheduler thread.
-  pjsip_rx_data* clone_rdata;
-  pj_status_t status = pjsip_rx_data_clone(rdata, 0, &clone_rdata);
-
-  if (status != PJ_SUCCESS)
-  {
-    // Failed to clone the message, so drop it.
-    LOG_ERROR("Failed to clone incoming message (%s)", PJUtils::pj_status_to_string(status).c_str());
-    return PJ_TRUE;
-  }
-
-  // Make sure the trail identifier is passed across.
-  set_trail(clone_rdata, get_trail(rdata));
-
-  // @TODO - need to think about back-pressure mechanisms.  For example,
-  // should we have a maximum depth of queue and drop messages after that?
-  // May be better to hold on to the message until the queue has space - this
-  // will force back pressure on the particular TCP connection.  Or should we
-  // have a queue per transport and round-robin them?
-
-  LOG_DEBUG("Queuing cloned received message %p for worker threads", clone_rdata);
-  qe.rdata = clone_rdata;
-
-  // Track the current queue size
-  queue_size_accumulator->accumulate(rx_msg_q.size());
-  rx_msg_q.push(qe);
-
-  // return TRUE to flag that we have absorbed the incoming message.
-  return PJ_TRUE;
-}
-
-
-static pj_status_t on_tx_msg(pjsip_tx_data* tdata)
-{
-  // Do logging.
-  local_log_tx_msg(tdata);
-  sas_log_tx_msg(tdata);
-
-  // Return success so the message gets transmitted.
-  return PJ_SUCCESS;
-}
-
 
 static void pjsip_log_handler(int level,
                               const char* data,
@@ -821,10 +478,22 @@ pj_status_t init_pjsip()
 
   status = register_custom_headers();
   PJ_ASSERT_RETURN(status == PJ_SUCCESS, status);
+  
+  return PJ_SUCCESS;
+}
 
-  // Enable deadlock detection on the message queue.
-  rx_msg_q.set_deadlock_threshold(MSG_Q_DEADLOCK_TIME);
+pj_status_t start_pjsip_thread()
+{
+  pj_status_t status = pj_thread_create(stack_data.pool, "pjsip", &pjsip_thread_func,
+			    NULL, 0, 0, &pjsip_thread);
+  if (status != PJ_SUCCESS)
+  {
+    LOG_ERROR("Error creating PJSIP thread, %s",
+	      PJUtils::pj_status_to_string(status).c_str());
+    return 1;
+  }
 
+  
   return PJ_SUCCESS;
 }
 
@@ -842,8 +511,6 @@ pj_status_t init_stack(const std::string& system_name,
                        const std::string& scscf_uri,
                        const std::string& alias_hosts,
                        SIPResolver* sipresolver,
-                       int num_pjsip_threads,
-                       int num_worker_threads,
                        int record_routing_model,
                        const int default_session_expires,
                        QuiescingManager *quiescing_mgr_arg,
@@ -855,11 +522,6 @@ pj_status_t init_stack(const std::string& system_name,
   pj_sockaddr addr_list[16];
   unsigned addr_cnt = PJ_ARRAY_SIZE(addr_list);
   unsigned i;
-
-  // Set up the vectors of threads.  The threads don't get created until
-  // start_stack is called.
-  pjsip_threads.resize(num_pjsip_threads);
-  worker_threads.resize(num_worker_threads);
 
   // Get ports and host names specified on options.  If local host was not
   // specified, use the host name returned by pj_gethostname.
@@ -970,10 +632,6 @@ pj_status_t init_stack(const std::string& system_name,
 
   // Initialise PJSIP and all the associated resources.
   status = init_pjsip();
-
-  // Register the stack module.
-  pjsip_endpt_register_module(stack_data.endpt, &mod_stack);
-  stack_data.module_id = mod_stack.id;
 
   // Initialize the PJUtils module.
   PJUtils::init();
@@ -1143,85 +801,14 @@ pj_status_t init_stack(const std::string& system_name,
 }
 
 
-pj_status_t start_stack()
+void term_pjsip_thread()
 {
-  pj_status_t status = PJ_SUCCESS;
-
-  quit_flag = PJ_FALSE;
-
-  // Create worker threads first as they take work from the PJSIP threads so
-  // need to be ready.
-  for (size_t ii = 0; ii < worker_threads.size(); ++ii)
-  {
-    pj_thread_t* thread;
-    status = pj_thread_create(stack_data.pool, "worker", &worker_thread,
-                              NULL, 0, 0, &thread);
-    if (status != PJ_SUCCESS)
-    {
-      LOG_ERROR("Error creating worker thread, %s",
-                PJUtils::pj_status_to_string(status).c_str());
-      return 1;
-    }
-    worker_threads[ii] = thread;
-  }
-
-  // Now create the PJSIP threads.
-  for (size_t ii = 0; ii < pjsip_threads.size(); ++ii)
-  {
-    pj_thread_t* thread;
-    status = pj_thread_create(stack_data.pool, "pjsip", &pjsip_thread,
-                              NULL, 0, 0, &thread);
-    if (status != PJ_SUCCESS)
-    {
-      LOG_ERROR("Error creating PJSIP thread, %s",
-                PJUtils::pj_status_to_string(status).c_str());
-      return 1;
-    }
-    pjsip_threads[ii] = thread;
-  }
-
-  return status;
-}
-
-void stop_stack()
-{
-  // Terminate the PJSIP threads and the worker threads to exit.  We kill
-  // the PJSIP threads first - if we killed the worker threads first the
-  // rx_msg_q will stop getting serviced so could fill up blocking
-  // PJSIP threads, causing a deadlock.
-
   // Set the quit flag to signal the PJSIP threads to exit, then wait
   // for them to exit.
   quit_flag = PJ_TRUE;
 
-  for (std::vector<pj_thread_t*>::iterator i = pjsip_threads.begin();
-       i != pjsip_threads.end();
-       ++i)
-  {
-    pj_thread_join(*i);
-  }
-
-  // Now it is safe to signal the worker threads to exit via the queue and to
-  // wait for them to terminate.
-  rx_msg_q.terminate();
-  for (std::vector<pj_thread_t*>::iterator i = worker_threads.begin();
-       i != worker_threads.end();
-       ++i)
-  {
-    pj_thread_join(*i);
-  }
+  pj_thread_join(pjsip_thread);
 }
-
-
-// Unregister all modules registered by the stack.  In particular, unregister
-// the transaction layer module, which terminates all transactions.
-void unregister_stack_modules(void)
-{
-  PJUtils::term();
-  pjsip_tsx_layer_destroy();
-  pjsip_endpt_unregister_module(stack_data.endpt, &mod_stack);
-}
-
 
 void term_pjsip()
 {
@@ -1251,9 +838,6 @@ void destroy_stack(void)
 
   delete connection_tracker;
   connection_tracker = NULL;
-
-  pjsip_threads.clear();
-  worker_threads.clear();
 
   SAS::term();
